Change default calculator to the NOAA (from USNO)
This relatively significant change replaces the USNO algorithm with the NOOA (Jean Meeus) one as the default calculator. The USNO algorithm does not account for leap years, and has been replaced by the USNO by newer and not publicly published algorithms (the code used in this API are based on the published Almanac from 1989 that was designed to be accurate for the years 1950-2050). Users who would like to use the USNO algorithm can easily continue doing so.