Use calculator's getUTCNoon in getSunTransit
This is as opposed to the previous behavior of returning getSunTransit(getSeaLevelSunrise(), getSeaLevelSunset()) that would be halfway between sunrise and sunset. If the NOOACalculator is used, it will now return proper astronomical noon, while the SunTimesCalculator that uses the UNO algorithm will continue using the time halfway between sunrise and sunset. See [The Definition of Chatzos](https://kosherjava.com/2020/07/02/definition-of-chatzos/) for additional information.
The ZmanimCalendar will have this as an option, defaulting to halfway between sunrise and sunset.

Also cleaned up some links.